# specs/002-auth-secure-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Web Application API with Authentication
  description: API for managing todo tasks with secure authentication in the full-stack web application
  version: 2.0.0  # Updated version to reflect authentication addition
servers:
  - url: http://localhost:8000/api
    description: Development server
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  description: User's password (minimum 8 characters)
                  minLength: 8
                  example: "securePassword123"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /auth/login:
    post:
      summary: Authenticate user and return JWT
      description: Authenticate user with email and password, return JWT access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  description: User's password
                  example: "securePassword123"
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /auth/me:
    get:
      summary: Get authenticated user info
      description: Retrieve information about the currently authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - insufficient permissions
        '500':
          description: Internal server error
  /tasks:
    get:
      summary: Get authenticated user's tasks
      description: Retrieve a list of all tasks for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of tasks for the authenticated user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - insufficient permissions
        '500':
          description: Internal server error
    post:
      summary: Create a new task for authenticated user
      description: Create a new todo task for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - insufficient permissions
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by its ID (must belong to authenticated user)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - task doesn't belong to authenticated user
        '404':
          description: Task not found
        '500':
          description: Internal server error
    put:
      summary: Update a task
      description: Update an existing task (must belong to authenticated user)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - task doesn't belong to authenticated user
        '400':
          description: Invalid input
        '404':
          description: Task not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a task
      description: Remove a task from the system (must belong to authenticated user)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - task doesn't belong to authenticated user
        '404':
          description: Task not found
        '500':
          description: Internal server error
  /tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Mark a task as complete or incomplete (must belong to authenticated user)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - completed
              properties:
                completed:
                  type: boolean
                  description: The new completion status for the task
                  example: true
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - task doesn't belong to authenticated user
        '400':
          description: Invalid input
        '404':
          description: Task not found
        '500':
          description: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - is_active
        - created_at
        - email_verified
      properties:
        id:
          type: integer
          description: Unique identifier for the user
          example: 1
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        is_active:
          type: boolean
          description: Whether the user account is active
          example: true
        email_verified:
          type: boolean
          description: Whether the user's email is verified
          example: true
        created_at:
          type: string
          format: date-time
          description: Timestamp when the user was created
          example: "2026-02-05T10:30:00Z"
        last_login:
          type: string
          format: date-time
          description: Timestamp of last successful login
          example: "2026-02-05T10:30:00Z"
    UserRegistrationResponse:
      type: object
      required:
        - id
        - email
        - created_at
        - email_verified
        - is_active
      properties:
        id:
          type: integer
          description: The new user's ID
          example: 1
        email:
          type: string
          format: email
          description: The new user's email
          example: "user@example.com"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the user was created
          example: "2026-02-05T10:30:00Z"
        email_verified:
          type: boolean
          description: Whether the user's email is verified
          example: false
        is_active:
          type: boolean
          description: Whether the user account is active
          example: true
    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        token_type:
          type: string
          description: Type of token (usually "bearer")
          example: "bearer"
        user:
          $ref: '#/components/schemas/User'
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - created_at
        - user_id
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: The title of the task
          example: "Buy groceries"
          maxLength: 255
        description:
          type: string
          description: Optional description of the task
          example: "Milk, eggs, bread, fruits"
          maxLength: 1000
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2026-02-05T10:30:00Z"
        user_id:
          type: integer
          description: ID of the user who owns this task
          example: 1
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The title of the task
          example: "Buy groceries"
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Optional description of the task
          example: "Milk, eggs, bread, fruits"
          maxLength: 1000
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: The new title of the task
          example: "Buy groceries"
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: New description of the task
          example: "Milk, eggs, bread, fruits"
          maxLength: 1000
        completed:
          type: boolean
          description: New completion status for the task
          example: false
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Invalid credentials"